+++ clang-tools-extra/clang-tidy/misrac/MisracTidyModule.cpp
@@ -38,6 +38,7 @@
 #include "Rule18p8Check.h"
 #include "Rule19p2Check.h"
 #include "Rule3p1Check.h"
+#include "Rule3p2Check.h"
 #include "Rule8p10Check.h"
 #include "Rule8p11Check.h"
 #include "Rule8p14Check.h"
@@ -112,6 +113,8 @@ public:
         "misrac-rule-19p2");
     CheckFactories.registerCheck<Rule3p1Check>(
         "misrac-rule-3p1");
+    CheckFactories.registerCheck<Rule3p2Check>(
+        "misrac-rule-3p2");
     CheckFactories.registerCheck<Rule8p10Check>(
         "misrac-rule-8p10");
     CheckFactories.registerCheck<Rule8p11Check>(